+++ src/common.h
@@ -108,6 +108,8 @@ namespace sharp {
   std::tuple<int, int> CalculateCrop(int const inWidth, int const inHeight,
     int const outWidth, int const outHeight, int const gravity);
 
+  int MaximumImageAlpha(VipsInterpretation interpretation);
+
 }  // namespace sharp
 
 #endif  // SRC_COMMON_H_