+++ src/pipeline.h
@@ -34,6 +34,7 @@ struct PipelineBaton {
   size_t overlayBufferInLength;
   int overlayGravity;
   bool overlayTile;
+  bool overlayCutout;
   int topOffsetPre;
   int leftOffsetPre;
   int widthPre;
@@ -99,6 +100,7 @@ struct PipelineBaton {
     overlayBufferInLength(0),
     overlayGravity(0),
     overlayTile(false),
+    overlayCutout(false),
     topOffsetPre(-1),
     topOffsetPost(-1),
     channels(0),